Consolidate protocol classes in protocols package
